Rename `proxy-init/pkg` to `proxy-init/internal`
* Move `proxy-init/pkg/iptables` to `proxy-init/internal/iptables`
* Move `proxy-init/pkg/ports` to `proxy-init/internal/util`
* Rename `IsValid` to `IsValidPort`
